+++ demo/bin/main/data.sql
@@ -2,4 +2,6 @@ INSERT INTO user (username, password, name, isAccountNonExpired, isAccountNonLoc
     VALUES ('user1', 'password', 'smallName1', 1, 1, 1, 1);
 
 INSERT INTO authority (username, authority_name)
+    VALUES ('user1', 'ADMIN');
+INSERT INTO authority (username, authority_name)
     VALUES ('user1', 'USER');
