+++ demo/bin/main/hello/mapper/LoginMapper.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="hello.UserMapper">
-  <select id="readUser" parameterType="String" resultType="hello.Member">
+<mapper namespace="hello.mapper.LoginMapper">
+  <select id="readUser" parameterType="String" resultType="hello.domain.Account">
     SELECT * FROM user WHERE username = #{username}
   </select>
   <select id="readAuthority" parameterType="String" resultType="String">
     SELECT authority_name FROM authority WHERE username = #{username}
   </select>
-  <insert id='insertUser' parameterType='hello.Member'>
+  <insert id="insertUser" parameterType="hello.domain.Account">
     INSERT INTO user(username, password, name, isAccountNonExpired, isAccountNonLocked, isCredentialsNonExpired, isEnabled)
     VALUES(#{username}, #{password}, #{name}, #{isAccountNonExpired}, #{isAccountNonLocked}, #{isCredentialsNonExpired}, #{isEnabled})
   </insert>
