+++ demo/src/main/java/hello/AppController.java
@@ -6,13 +6,13 @@ import org.springframework.stereotype.Controller;
 
 @Controller
 public class AppController {
-    @RequestMapping("/home")
-    public String index() {
-        return "index";
-    }
+    // @RequestMapping("/home")
+    // public String index() {
+    //     return "index";
+    // }
 
-    @RequestMapping("/login")
-    public String login() {
-        return "Login/login";
-    }
+    // @RequestMapping("/login")
+    // public String login() {
+    //     return "Login/login";
+    // }
 }
