--- demo/src/main/java/hello/Controller/LoginController.java
@@ -1,12 +1,0 @@
-package hello.Controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class LoginController{
-    // @RequestMapping("/login")
-    // public String labels(){
-    //     return "Login/login";
-    // }
-}
