+++ demo/src/main/java/hello/MyNoPasswordEnconder.java
@@ -1,15 +1,15 @@
-package hello;
+// package hello;
 
-import org.springframework.security.crypto.password.PasswordEncoder;
+// import org.springframework.security.crypto.password.PasswordEncoder;
 
-public class MyNoPasswordEnconder implements PasswordEncoder {
-    @Override
-    public String encode(CharSequence charSequence) {
-        return charSequence.toString();
-    }
+// public class MyNoPasswordEnconder implements PasswordEncoder {
+//     @Override
+//     public String encode(CharSequence charSequence) {
+//         return charSequence.toString();
+//     }
 
-    @Override
-    public boolean matches(CharSequence charSequence, String s) {
-        return charSequence.toString().equals(s);
-    }
-} 
+//     @Override
+//     public boolean matches(CharSequence charSequence, String s) {
+//         return charSequence.toString().equals(s);
+//     }
+// }
