+++ demo/src/main/java/hello/SecurityMember.java
@@ -1,12 +1,14 @@
-package hello;
+// package hello;
 
-import org.springframework.security.core.userdetails.User;
+// import org.springframework.security.core.userdetails.User;
 
-public class SecurityMember extends User {
+// public class SecurityMember extends User {
 
-    private static final long serialVersionUID = 1L;
+//     private static final long serialVersionUID = 1L;
 
-    public SecurityMember(Member member) {
-        super(member.getUsername(), member.getPassword(), member.getAuthorities());
-    }
-} 
+//     public SecurityMember(Member member) {
+//         super(member.getUsername(), member.getPassword(), member.getAuthorities());
+//     }
+// } 
+
+//사용자의 Authority정보를 조회해서 세션을 유지시켜주는곳
