+++ demo/src/main/java/hello/UserMapper.java
@@ -1,11 +1,11 @@
-package hello;
+// package hello;
 
-import java.util.List;
+// import java.util.List;
 
-import hello.Member;
+// import hello.Member;
 
-public interface UserMapper {
-    public Member readUser(String username);
-    public List<String> readAuthority(String username);
-    public void insertUser(Member user);
-}
+// public interface UserMapper {
+//     public Member readUser(String username);
+//     public List<String> readAuthority(String username);
+//     public void insertUser(Member user);
+// }
