+++ demo/src/main/java/hello/controller/HomeController.java
@@ -1,4 +1,4 @@
-package hello.Controller;
+package hello.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
