+++ demo/src/main/java/hello/controller/LabelsController.java
@@ -1,0 +1,12 @@
+package hello.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+public class LabelsController{
+    @RequestMapping("/labels")
+    public String labels(){
+        return "Labels/labels";
+    }
+}
