+++ demo/src/main/java/hello/controller/SignUpController.java
@@ -1,6 +1,5 @@
-package hello.Controller;
+package hello.controller;
 
-import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,20 +7,13 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 
-import hello.Member;
-import hello.UserMapper;
+import hello.domain.Account;
 import hello.service.SignUpService;
 
 @Controller
 public class SignUpController{
-    // @Resource(name="hello.service.SignUpService")
-    // SignUpService signUpService;
-
-    @Autowired
-    private UserMapper userMapper;
+    SignUpService signUpService;
 
 	@Autowired
 	PasswordEncoder passwordEncoder;
@@ -33,11 +25,12 @@ public class SignUpController{
 
     @PostMapping("/signUpOk")
     public String signUpOk(HttpServletRequest request){
-        Member member = new Member();
-        member.setUsername(request.getParameter("username"));
-        member.setName(request.getParameter("name"));
-        member.setPassword(passwordEncoder.encode(request.getParameter("password")));
-        userMapper.insertUser(member);
+        Account account = new Account();
+        account.setUsername(request.getParameter("username"));
+        account.setName(request.getParameter("name"));
+        account.setPassword(passwordEncoder.encode(request.getParameter("password")));
+
+        signUpService.signUpOkService(account);
         System.out.println("회원가입완료");
         return "/index";
     }
