+++ demo/src/main/java/hello/domain/Account.java
@@ -1,0 +1,23 @@
+package hello.domain;
+
+import java.util.Collection;
+
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import lombok.Data;
+
+@Data
+public class Account implements UserDetails{
+    private Long id;
+    private String username;   
+    private String password;
+    private String name;
+    private boolean isAccountNonExpired;
+    private boolean isAccountNonLocked;
+    private boolean isCredentialsNonExpired;
+    private boolean isEnabled;
+
+    private Collection<? extends GrantedAuthority> authorities;
+}
+
