+++ demo/src/main/java/hello/mapper/LoginMapper.java
@@ -1,0 +1,11 @@
+package hello.mapper;
+
+import java.util.List;
+
+import hello.domain.Account;
+
+public interface LoginMapper {
+    public Account readUser(String username);
+    public List<String> readAuthority(String username);
+    public void insertUser(Account user);
+}
