+++ demo/src/main/java/hello/service/login/CustomUserDetailsService.java
@@ -1,4 +1,4 @@
-package hello;
+package hello.service.login;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -11,26 +11,32 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import hello.Member;
-import hello.SecurityMember;
-import hello.UserMapper;
+import hello.domain.Account;
+import hello.mapper.LoginMapper;
+import hello.service.login.SecurityMember;
 
 @Service
+//UserDetailsService 인터페이스는 DB에서 유저 정보를 가져오는 역할을 한다
 public class CustomUserDetailsService implements UserDetailsService{
 
     private static final String ROLE_PREFIX = "ROLE_";
 
     @Autowired
-    UserMapper userMapper;
+    LoginMapper loginMapper;
 
+    /**
+     * UserDetailsService 인터페이스에는 DB에서 유저 정보를 불러오는 중요한 메소드가 존재한다. 
+     * 바로 loadUserByUsername() 메소드이다. 
+     * 이 메소드에서 유저 정보를 불러오는 작업을 하면 된다.
+     */
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
 
-        Member member = userMapper.readUser(username);
-        if(member != null) {
-            member.setAuthorities(makeGrantedAuthority(userMapper.readAuthority(username)));
+        Account account = loginMapper.readUser(username);
+        if(account != null) {
+            account.setAuthorities(makeGrantedAuthority(loginMapper.readAuthority(username)));
         }
-        return new SecurityMember(member);
+        return new SecurityMember(account);
     }
 
     private static List<GrantedAuthority> makeGrantedAuthority(List<String> roles){
