+++ demo/src/main/java/hello/service/login/CustomWebSecurityConfigurerAdapter.java
@@ -1,0 +1,61 @@
+package hello.service.login;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.builders.WebSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+
+import hello.service.login.CustomUserDetailsService;
+import hello.service.login.CustomLoginSuccessHandler;
+ 
+@EnableWebSecurity     
+public class CustomWebSecurityConfigurerAdapter extends WebSecurityConfigurerAdapter {
+    
+    @Autowired
+    CustomUserDetailsService customUserDetailsService;
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    @Bean
+    public AuthenticationSuccessHandler successHandler() {
+        return new CustomLoginSuccessHandler("/issue_list");
+    }
+
+    @Override           
+    public void configure(WebSecurity web) throws Exception {
+        web.ignoring().antMatchers(
+            "/resources/**"
+        );
+    }
+    
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http
+            .csrf().disable()
+            .authorizeRequests()
+            .antMatchers("/", "/home", "/signUp").permitAll()
+            .anyRequest().authenticated();    
+        http.formLogin()
+            .loginPage("/login")
+            .successHandler(successHandler())
+            .permitAll();        
+        http.logout()
+            .logoutUrl("/logout")
+            .logoutSuccessUrl("/")
+            .permitAll();
+    }
+ 
+    @Override
+    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+        auth.userDetailsService(customUserDetailsService).passwordEncoder(passwordEncoder());
+    }
+}
