+++ demo/src/main/java/hello/service/login/MyNoPasswordEnconder.java
@@ -1,0 +1,15 @@
+package hello.service.login;
+
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+public class MyNoPasswordEnconder implements PasswordEncoder {
+    @Override
+    public String encode(CharSequence charSequence) {
+        return charSequence.toString();
+    }
+
+    @Override
+    public boolean matches(CharSequence charSequence, String s) {
+        return charSequence.toString().equals(s);
+    }
+}
