--- demo/src/main/webapp/WEB-INF/jsp/hello.jsp
@@ -1,13 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=utf-8"
-         pageEncoding="utf-8" isELIgnored="false"%>
-<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
-
-<html>
-<head>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <title>Spring Boot Application with JSP</title>
-</head>
-<body>
-    Hello, Spring Boot App
-</body>
-</html>
